miner/miner-base/pyproject.toml
[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "pytest-asyncio>=1.2.0",
  "pytest-xdist>=3.8.0",
  "pytest>=8.3.0",
  "ruff>=0.13.0"
]

[project]
name = "miner-base"
version = "0.1.0"
description = "Common utilities and meta-package for Pearl miner"
requires-python = ">=3.8"
dependencies = [
  "blake3>=1.0.7",
  "miner-utils",
  "numpy>=1.20.0",
  "pearl-gateway",
  "py-pearl-mining",
  "pydantic-settings>=2.12.0",
  "torch==2.9.1",
]

